# Specialized algorithm for different ranges of values.