For a one-liner this code sure is getting long.

I worked out the performance issue I was having - LLVM was getting really confused by a by-value `sum` taking a `[T; 4]`. Changing that to a reference fixed things.

I've tuned the specific constants for my machine. Verification that I haven't caused a regression would be helpful. When doing so you might notice there are a *lot* more benchmarks. This is getting unwieldy, but they've been very useful for me.

Fixes #10.