@llogiq 2016-12-03
The errors seem to be in the benchmark code. Perhaps an incomplete commit?