@llogiq 2016-12-03
We could include a build script to determine the optimum thresholds for the given machine.